1066 - Separated converter to package
