+++ src/main/java/algorithms/curated170/medium/CampusBikes2.java
@@ -3,6 +3,7 @@ package algorithms.curated170.medium;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import algorithms.util.leetcode.LCArrayConverter;
 
 public class CampusBikes2 {
 
@@ -45,45 +46,10 @@ public class CampusBikes2 {
       public static void main(String[] args) {
 
             var solution = new CampusBikes2();
-            int[][] workers = leetCodeIntegerGridConverter("[[1,1],[1,3],[1,5],[4,5]]");
-            int[][] bikes = leetCodeIntegerGridConverter("[[4,2],[2,2],[3,5],[5,1]]");
+            int[][] workers = LCArrayConverter.leetCodeIntegerGridConverter("[[1,1],[1,3],[1,5],[4,5]]");
+            int[][] bikes = LCArrayConverter.leetCodeIntegerGridConverter("[[4,2],[2,2],[3,5],[5,1]]");
 
             int sol = solution.assignBikes(workers, bikes);
             System.out.println(sol);
-      }
-
-      public static int[][] leetCodeIntegerGridConverter(String arr) {
-            List<List<Integer>> list = new ArrayList<>();
-
-            List<Integer> currAr = new ArrayList<>();
-            char prev = '#';
-            int num = 0;
-            boolean inNum = false;
-            for (char c : arr.toCharArray()) {
-                  if (!Character.isDigit(c) && inNum) {
-                        currAr.add(num);
-                        num = 0;
-                        inNum = false;
-                  }
-                  if (c == '[') {
-                        currAr = new ArrayList<Integer>();
-                  } else if (Character.isDigit(c)) {
-                        num = num * 10 + Character.getNumericValue(c);
-                        inNum = true;
-                  } else if (c == ']' && prev != c) {
-                        list.add(currAr);
-                  }
-                  prev = c;
-            }
-            System.out.println(list);
-            Integer[][] arrayBase = list.stream().map(l -> l.stream().toArray(Integer[]::new))
-                        .toArray(Integer[][]::new);
-            int[][] ret = new int[arrayBase.length][arrayBase[0].length];
-            for (int i = 0; i < arrayBase.length; i++) {
-                  for (int j = 0; j < arrayBase[0].length; j++) {
-                        ret[i][j] = arrayBase[i][j];
-                  }
-            }
-            return ret;
       }
 }
